app/tasks/handlers/faq.py
"""FAQ 导入任务处理器

对齐 WeKnora99 的 FAQ 导入功能 (TypeFAQImport)
"""

from sqlalchemy.ext.asyncio import AsyncSession

from app.observability.logging import get_logger
from app.tasks.handlers.base import task_context

logger = get_logger(__name__)


async def process_faq_import(
    celery_task,
    payload: dict,
    tenant_id: int,
) -> dict:
    """处理 FAQ 导入任务

    Args:
        celery_task: Celery 任务实例
        payload: 任务参数
        tenant_id: 租户 ID

    Returns:
        处理结果
    """
    task_id = payload.get("task_id")
    kb_id = payload.get("kb_id")
    entries = payload.get("entries", [])
    dry_run = payload.get("dry_run", False)

    from app.infra.database import get_session_factory

    session_factory = get_session_factory()

    async with session_factory() as session:
        async with task_context(celery_task, session, task_id, tenant_id) as handler:
            # 创建任务记录
            await handler.create_task(
                task_type="faq:import",
                payload=payload,
                title=f"FAQ 导入 ({'Dry Run' if dry_run else '正式导入'})",
                business_id=kb_id,
                business_type="knowledge_base",
                total_items=len(entries),
            )

            await handler.mark_started()

            try:
                # 如果是 dry run，只验证不导入
                if dry_run:
                    await handler.update_progress(50, "验证 FAQ 条目")

                    # TODO: 实现验证逻辑
                    # validation_results = await _validate_faq_entries(entries)
                    # valid_count = sum(1 for r in validation_results if r.is_valid)

                    await handler.update_progress(100, "验证完成")

                    result = {
                        "status": "completed",
                        "dry_run": True,
                        "total_entries": len(entries),
                        # "valid_count": valid_count,
                    }

                    await handler.mark_completed(result)
                    return result

                # 正式导入
                await handler.update_progress(10, "开始导入 FAQ 条目")

                # TODO: 实现导入逻辑
                # for i, entry in enumerate(entries):
                #     await handler.update_progress(
                #         int((i / len(entries)) * 100),
                #         f"导入: {entry.get('question', '')[:30]}",
                #         processed_items=i + 1,
                #     )
                #     await _import_faq_entry(session, entry, kb_id, tenant_id)

                result = {
                    "status": "completed",
                    "dry_run": False,
                    "total_entries": len(entries),
                    "kb_id": kb_id,
                }

                await handler.mark_completed(result)
                return result

            except Exception as e:
                import traceback

                await handler.mark_failed(str(e), traceback.format_exc())
                raise
